requirements.txt
tensorflow-cpu==2.13.0
streamlit
Pillow
numpy
opencv-python-headless